+++ src/components/splitButton/SplitButton.tsx
@@ -9,6 +9,7 @@ import { BoxProps } from '../box/Box';
 import { GenericComponent } from '../../@types/types';
 import isReactElement from '../utils/is-react-element';
 import { SIZES } from '../../constants';
+import theme from './theme.css';
 
 interface SplitButtonProps extends Omit<BoxProps, 'children' | 'size'> {
   /** The MenuItems we pass to our component. */
@@ -23,6 +24,17 @@ interface SplitButtonProps extends Omit<BoxProps, 'children' | 'size'> {
   onSecondaryButtonClick?: (event: MouseEvent<HTMLElement>) => void;
   /** If true, component will be disabled. */
   disabled?: boolean;
+  /** If true, component will show a loading spinner instead of label. */
+  processing?: boolean;
+  /** Overwrites for the popover */
+  popoverProps?: {
+    /** The preferred direction in which the Popover is rendered, is overridden with the opposite or adjacent direction if the Popover cannot be entirely displayed in the current direction. */
+    direction?: 'north' | 'south';
+    /** If true, the Popover stretches to fit its content vertically */
+    fullHeight?: boolean;
+    /** The z-index of the Popover */
+    zIndex?: number;
+  };
 }
 
 const SplitButton: GenericComponent<SplitButtonProps> = ({
@@ -32,6 +44,8 @@ const SplitButton: GenericComponent<SplitButtonProps> = ({
   onButtonClick,
   onSecondaryButtonClick,
   disabled,
+  processing,
+  popoverProps,
   ...others
 }) => {
   const [popoverActive, setPopoverActive] = useState<boolean>(false);
@@ -72,16 +86,26 @@ const SplitButton: GenericComponent<SplitButtonProps> = ({
 
   return (
     <Box display="flex" justifyContent="center" {...boxProps} data-teamleader-ui="split-menu">
-      <ButtonGroup segmented>
-        <Button label={buttonLabel} level={level} size={size} disabled={disabled} onClick={handleMainButtonClick} />
+      {processing ? (
         <Button
-          icon={<IconChevronDownSmallOutline />}
+          label={buttonLabel}
           level={level}
           size={size}
-          disabled={disabled}
-          onClick={handleSecondaryButtonClick}
+          processing
+          className={theme[`split-button-${size}--processing`]}
         />
-      </ButtonGroup>
+      ) : (
+        <ButtonGroup segmented>
+          <Button label={buttonLabel} level={level} size={size} disabled={disabled} onClick={handleMainButtonClick} />
+          <Button
+            icon={<IconChevronDownSmallOutline />}
+            level={level}
+            size={size}
+            disabled={disabled}
+            onClick={handleSecondaryButtonClick}
+          />
+        </ButtonGroup>
+      )}
       <Popover
         active={popoverActive}
         anchorEl={popoverAnchorEl?.current}
@@ -90,6 +114,7 @@ const SplitButton: GenericComponent<SplitButtonProps> = ({
         onEscKeyDown={handleCloseClick}
         onOverlayClick={handleCloseClick}
         position="start"
+        {...popoverProps}
       >
         <Menu>
           {React.Children.map(children, (child) => {